Refactor the code structure to make it easy to document
